requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.1
pandas==2.1.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
newspaper3k==0.2.8
requests==2.31.0
feedparser==6.0.10
lxml==4.9.3
python-dateutil==2.8.2
nltk==3.8.1
tqdm==4.66.1
fake-useragent==1.4.0
httpx==0.25.2
pydantic-settings==2.1.0

# LLM and AI dependencies
openai==1.3.0
tiktoken==0.5.2
langchain==0.0.335
langchain-openai==0.0.2
